requirements.txt
langgraph
langchain
langchain-openai
langchain-community
langchain-chroma
chromadb
python-frontmatter
openai
faiss-cpu
fastapi
uvicorn
python-multipart
markdown-it-py
gradio>=4,<5
huggingface-hub<1.0
python-dotenv>=1.0.0
pdfplumber
python-docx
